requirements.txt
alpaca-trade-api
pandas
numpy
python-dotenv
sanpy
pycoingecko
